# passwords and glob ffmpeg in win