@Theelgirl 2020-10-09
That's super interesting! Did you make any benchmarks for speed?