@Theelgirl 2020-10-09
Also, when I tested it, it said "unrecognized argument: test" when I passed ``-p "test"``. I see that passing -p alone makes it ask you for the password. I suggest that you make it let you pass the password in the command line by removing the action parameter, and if you do that, then it doesn't ask for your input.