@dobrosketchkun 2020-10-09
Do you mean just type your password as plain text? It's not very secure.

>A short help description may be nice