@dobrosketchkun 2020-10-09
It looks like a nice compromise!

>If you'd like, I can add this to my optimization commit.

It'll be nice as long as it'll not be confusing for AlfredoSequeida and if it is tested with full pasword code modifications.