@Theelgirl 2020-10-09
I'm getting this when I try to decrypt videos, do you know what it means?
``gzip.BadGzipFile: Not a gzipped file (b'\xff\xd8')``